backend/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
spotipy==2.23.0
discogs-client==2.3.0
anthropic==0.18.1
mailchimp3==3.0.19
flask-cors==4.0.0
python-jose==3.3.0
gunicorn==21.2.0 